modeltests/requirements.txt
matplotlib==3.5.1
numpy==1.22.2
onnx==1.10.2
onnxruntime==1.10.0
pyfunctional==1.4.3
tensorflow==2.8.0
torch==1.10.2
transformers==4.16.2
